// IdentityScoreJult28/ClientApp/app/Interfaces/participant.interface.ts
import { socialmediaprofile } from '../Interfaces/socialmediaprofile.interface';

export interface participant {
    FirstName?: string;
    LastName?: string;
    EmailAddress?: string;
    Password?: string;
    phcode?: string;
    confirmPassword?: string;
    PhoneNumber?: string;
    City?: string;
    State?: string;
    Zip?: number;
    GenderIdentity?: string;
    SexualOrientation?: string;
    Race?: string;
    Hispanic?: string;
    Age?: number;
    Date_of_Birth?: Date;
    othergender?: string;
    otherSex?: string;
    otherRace?: string;
    YearofBirth?: number;
    MonthofBirth?: string;
    captcha?: string;
    }
    /* IPAddress?: string, Browser?: string, OS?: string, DeviceType?: string, AccountCreated?: Date*/